// frontend/src/lib/schemas/booking.ts
import * as z from "zod"

export const BookingSchema = z.object({
  firstName: z.string().min(1, "First name is required"),
  lastName: z.string().min(1, "Last name is required"),
  email: z.string().email("Invalid email address"),
  subject: z.string().min(1, "Subject is required"),
  message: z.string().min(1, "Message is required"),
})

export type BookingFormData = z.infer<typeof BookingSchema>;